backend/requirements.txt
Django==3.2
djangorestframework==3.14.0
djoser==2.2.0